Update `observation_type` and `base_sgd`
