+++ avalanche/training/templates/common_templates.py
@@ -328,7 +328,7 @@ class OnlineSupervisedTemplate(OnlineObservation, SupervisedProblem, SGDUpdate,
             optimizer=optimizer,
             criterion=criterion,
             train_mb_size=train_mb_size,
-            train_epochs=1,
+            train_epochs=train_passes,
             eval_mb_size=eval_mb_size,
             device=device,
             plugins=plugins,
